Fix: Completing a task was showing Next task as complete wrongly
To reproduce:
Task mode, complete a task. Next task is returned correctly, but status shows in UI as complete. Basically using the old status. This doesn't make sense in this context since expectation was file changed.

There could be some edge case on the last file item not updating but for now that will be an edge case. And in that case needs more refactor to make more readable and useful.