Add label context menu & UI/UX improvements (#1407)
* Update README.md

* Clean

* Add label select change from context menu

* Clean

-> This way shows cleanly over whole context menu instead of icons

* clean

* Remove legacy function box_update_position

This is now handled in current_instance it appears.
(Function is not used anywhere)

* remove unneeded condition

reminder that instance_hover_index_locked can be 0

* clean

* clean

* clean name

* Fix timing issue setting color on box

-> Timing issue where label_file_id was changed, but this function is firing before other function can fire
-> Added simple cache to maintain speed, however even without this it was showing 0 MS time.

There may be other places this needs to be reviewed

Also not really clear if "caching" this here is actually that optimal or if the function difference makes sense here. Seems a bit too tied together since you would need ot know to use set_color_from_label() and wait before calling get_color_from_instance() if it has been changed.

* Update InstanceImage2D.ts

* Improve default speed

this is use selectable
but can slow loading if it's loading additional pages

* clean commented code

* Set default width shorter in performance mode

* Align with new annotation_ui_context paradigm

Before this was in the function,
but since it's an object it's not watching it properly
the object is updating, so can add specific property in this context

appears to work better,
need to define better standards around this reactivity with objects here

* Clean

* Move to proper header

Makes more sense for this to be part of the global header

* remove unused code

* Align with new Sidebar paradigm

Now the $refs don't exit.
Align with existing focus() and focus_instance_show_all() show all functions. All this is missing is emitting an event so that detail list can know about the focus too

* Update context_menu.vue

* Upgrade to issues_ui_manager paradigm

* clean old var