Big refactor on media queue and add copy queue
-> Move all ProcessMediaQueueManager functions into that existing class and out of pure functional from legacy file.
-> Simplify all functions including: --> Create new router() function
-> Pull out other functions
It makes so much more sense to have it all in this one class
-> Remove confusing double while true loop
-> Remove confusing remote_queue_sleep_time logic. This added a huge amount of confusion, but very little value. Assumptions around guessing at how often we should check here made little sense. There may be valid reasons to change this time, but the ones implemented were not it.

This now avoids confusing imports and makes process media much cleaner.

Note that there isn't a one : one correlation between the local and remote queues, so for now there could be confusion there if one queue is over limit which is a weakspot.

For now, setting a queue limit to be None will have it "non-blocking" for the remote loop. Another option would be to do different queries per media type. But now that is more possible.

Everything should be in PROCESSING_INPUT_LIST so removed other lines