Refactor Process Media Queues (#1437)
* clean

* Clean

* clean

* refactor

confusing having this only at function level

* Add support for time_last_updated

Not sure if we want this to be a more "general" thing, e.g.
all class attrs are available.

Also kind of awkward in way we are updating https://diffgram.readme.io/docs/query-syntax-reference

Want to more generally expose these fields

* Add ignore newline (\n)

Otherwise in UI explorer context can randomly break without clear reason. e.g. UI adding in file.created_time >= \"2023-06-23\"\n"
At the moment new line is not a support thing,

Could also strip this in frontend but this way more durable or API context too.

* clarify

* clean unused vars

* Clean unused var and old comments

Temp dir comments don't make sense here anymore since no temp dir created in this function

* Big refactor on media queue and add copy queue

-> Move all ProcessMediaQueueManager functions into that existing class and out of pure functional from legacy file.
-> Simplify all functions including: --> Create new router() function
-> Pull out other functions
It makes so much more sense to have it all in this one class
-> Remove confusing double while true loop
-> Remove confusing remote_queue_sleep_time logic. This added a huge amount of confusion, but very little value. Assumptions around guessing at how often we should check here made little sense. There may be valid reasons to change this time, but the ones implemented were not it.

This now avoids confusing imports and makes process media much cleaner.

Note that there isn't a one : one correlation between the local and remote queues, so for now there could be confusion there if one queue is over limit which is a weakspot.

For now, setting a queue limit to be None will have it "non-blocking" for the remote loop. Another option would be to do different queries per media type. But now that is more possible.

Everything should be in PROCESSING_INPUT_LIST so removed other lines

* Better isolate file operations

* remove unused import

* Update to class and router

* Clean old comments

* remove in context of confusing imports

it's confusing to instantiate the thing here
we already have this in main, so it's doubling up on it too potentially

* align with rest of pattern

importing object only not module
and then using class directly

* align standard logger

* Refactor into own file

Help to clean files and improve imports

* Update prioritized_item.py

* Move to shared

Context of using in shared/packet.py
In general these abstract definitions work better in shared

* Create new single file for starting queue

Helps make it super clear where it is importing it or not

* remove duplicate import

* this is expected correct case, no need to log

* Seperate upload class and API

there are assumptions around how sub-imports are used in shared context
by separating this out, we can more cleaning use Upload class in shared context without affecting flask route

* Little clean ups and refactor fixes

* Move to dedicated file for concrete instance

For more durable usage in shared context we should not assume to instantiate the class in the class definition!

Now it has dedicated file for concrete instance

Also align shared abstract definition of PrioritizedItem into shared folder

Also process_media_queue_manager is only needed in walrus context on packets only import it there.

* Fix legacy circular import assumptions

`New_video` is part of legacy assumptions
Align with rest of pattern in file for handling that

And align with abstract shared definition of PrioritizedItem

* Update process_media_queue_manager.py